requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
gunicorn
psycopg2-binary
coverage==7.8.0
crispy-bootstrap5==2025.4
cryptography==44.0.2
Django==5.2
django-allauth==65.7.0
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
jwt==1.3.1
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise
social-auth-app-django
python-decouple
selenium